+++ CrypMulator/CrypMulator/Domain/OldFavoriteCoinRepository.swift
@@ -5,6 +5,7 @@
 //  Created by 김태형 on 3/9/25.
 //
 
+// TODO: - 추후 제거
 import Foundation
 import RealmSwift
