+++ CrypMulator/CrypMulator/Domain/Repository/ChargeRepositoryProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  ChargeRepositoryProtocol.swift
+//  CrypMulator
+//
+//  Created by 김태형 on 4/2/25.
+//
+
+import Foundation
+
+protocol ChargeRepositoryProtocol: AnyObject {
+    func getFileURL()
+    func saveCharge(amount: Decimal)
+}
